requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
pycryptodome==3.18.0
Werkzeug==2.3.7